api_capstone/requirements.txt
# --------------------
# Core Frameworks
# --------------------
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.2

# --------------------
# ORM & Database
# --------------------
SQLAlchemy==2.0.41
PyMySQL==1.1.1
alembic==1.16.1

# --------------------
# Auth & Security
# --------------------
bcrypt==4.3.0
passlib==1.7.4
python-jose==3.4.0
email_validator==2.2.0

# --------------------
# Pydantic & Validation
# --------------------
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
pydantic-extra-types==2.10.4
annotated-types==0.7.0

# --------------------
# API & HTTP Tools
# --------------------
httpx==0.28.1
requests==2.32.3
python-multipart==0.0.20
aiohttp==3.11.18
aiohappyeyeballs==2.6.1
httptools==0.6.4
websockets==15.0.1

# --------------------
# Logging & CLI
# --------------------
rich==14.0.0
rich-toolkit==0.14.6
click==8.1.8
typer==0.15.4
shellingham==1.5.4

# --------------------
# Testing
# --------------------
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-dotenv==0.5.2
pytest-mock==3.14.0
coverage==7.8.2

# --------------------
# Environment & Config
# --------------------
python-dotenv==1.1.0
PyYAML==6.0.2

# --------------------
# Linting & Static Analysis
# --------------------
pylint==3.3.7
isort==6.0.1
mccabe==0.7.0
astroid==3.3.10
typeguard==4.4.2

# --------------------
# ML/DL & Transformers
# --------------------

# --------------------
# GPU / CUDA (cu12)
# --------------------
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105

# --------------------
# Datasets & NLP
# --------------------
datasets==3.6.0
pandas==2.2.3
numpy==1.26.4
pyarrow==20.0.0
multiprocess==0.70.16
dill==0.3.8
regex==2024.11.6
orjson==3.10.18
ujson==5.10.0

# --------------------
# Cloudinary & Media
# --------------------
cloudinary==1.44.0
pillow==11.2.1

# --------------------
# FastAPI CLI & Tools
# --------------------
fastapi-cli==0.0.7
docstring_parser==0.16

# --------------------
# Utils & Internal
# --------------------
mediatr==1.3.2
propcache==0.3.1
tzdata==2025.2
xxhash==3.5.0
packaging==25.0
fsspec==2025.3.0
filelock==3.18.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
sniffio==1.3.1
platformdirs==4.3.8
zipp==3.22.0
greenlet==3.2.2
importlib_metadata==8.7.0
iniconfig==2.1.0
mdurl==0.1.2
markdown-it-py==3.0.0
frozenlist==1.6.0
yarl==1.20.0

# --------------------
# Cryptography & Security
# --------------------
cffi==1.17.1
cryptography==45.0.2
pycparser==2.22
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.4.8

